feat: 用户取消操作时显示提示消息而不是报错
resolves #24

更多修改:
修改目录结构，减少不必要的重复
修改安装程序图标
