+++ pack.iss
@@ -7,10 +7,9 @@ VersionInfoCopyright=Copyright (c) 鸭鸭「カモ」
 AppPublisherURL=https://duckduckstudio.github.io/yazicbs.github.io/
 DefaultDirName={autopf}\Fufu_Dev_Tools
 DefaultGroupName=芙芙工具箱开发工具包
-UninstallDisplayIcon={app}\Script\ffdev.exe
 OutputDir=Release
 OutputBaseFilename=Fufu_Dev_Tools_Setup_develop
-SetupIconFile=Fufu_Tools\src\ico.ico
+SetupIconFile=Fufu_Tools\Installer\Fufu_Tools_Setup_ico.ico
 LicenseFile=LICENSE-GBK
 Compression=lzma2
 SolidCompression=yes
@@ -21,12 +20,7 @@ Name: "chinesesimplified"; MessagesFile: "compiler:Languages\ChineseSimplified.i
 Name: "japanese"; MessagesFile: "compiler:Languages\Japanese.isl"
 
 [Files]
-Source: "开发工具-打包-done\*"; DestDir: "{app}"
-Source: "开发工具-打包-done\git\*"; DestDir: "{app}\git"
-Source: "开发工具-打包-done\Script\*"; DestDir: "{app}\Script"
-Source: "开发工具-打包-done\代码校对\*"; DestDir: "{app}\代码校对"
-Source: "开发工具-打包-done\生成工具\*"; DestDir: "{app}\生成工具"
-Source: "开发工具-打包-done\统计\*"; DestDir: "{app}\统计"
+Source: "src-pack\*"; DestDir: "{app}" Flags: recursesubdirs
 
 [Icons]
 Name: "{group}\芙芙工具箱开发工具包"; Filename: "{app}\Script\ffdev.exe"
