+++ src/代码校对/尾随空格.py
@@ -1,5 +1,6 @@
 import os
 import re
+import sys
 import argparse
 import chardet
 from colorama import init, Fore
@@ -51,16 +52,24 @@ def main(directory, file_extensions=None):
     if directory:
         check_files(directory, file_extensions)
         print(f"{Fore.GREEN}✓{Fore.RESET} 尾随空格检查完成。")
+        return 0
+    else:
+        print(f"{Fore.RED}✕{Fore.RESET} 请提供要检查的目录")
+        return 1
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description="检查指定目录下的文件是否存在尾随空格")
-    parser.add_argument("--file_extensions", dest="file_extensions", type=str, help="要检查的文件扩展名列表，以 , 分隔")
-    parser.add_argument("--dir", dest="directory", required=True, help="要检查的目录路径")
-    args = parser.parse_args()
+    try:
+        parser = argparse.ArgumentParser(description="检查指定目录下的文件是否存在尾随空格")
+        parser.add_argument("--file_extensions", dest="file_extensions", type=str, help="要检查的文件扩展名列表，以 , 分隔")
+        parser.add_argument("--dir", dest="directory", required=True, help="要检查的目录路径")
+        args = parser.parse_args()
 
-    if args.file_extensions:
-        # 将逗号分隔的字符串转换为列表
-        file_extensions = args.file_extensions.split(',') if args.file_extensions else []
-        main(args.directory, file_extensions)
-    else:
-        main(args.directory)
+        if args.file_extensions:
+            # 将逗号分隔的字符串转换为列表
+            file_extensions = args.file_extensions.split(',') if args.file_extensions else []
+            main(args.directory, file_extensions)
+        else:
+            main(args.directory)
+    except KeyboardInterrupt:
+        print(f"{Fore.RED}✕{Fore.RESET} 用户已取消操作")
+        sys.exit(2)
