--- 开发工具-源码/config.ini
@@ -1,7 +1,0 @@
-[Check_trailing_space]
-FileExtensions = html,css,js,py,pyw,ini,txt,bat,json,c
-# 仅为示例，用于设定总调用程序中检测尾随空格时只检测的文件格式。
-
-[Statistics]
-CodeFileExtensions = cpp,cxx,py,pyw,js,html,css,java,e
-# 仅为示例，可自行增减，用于设定统计代码总行数工具 --code-file-formats 参数指定的文件格式。
